STAR-515 Add guardrail for SASI and STAR-512 Add guardrails for total and per-table indexes for SAI (#140)
* STAR-512 Add guardrails for total and per-table indexes for SAI


* STAR-515 Add guardrail for SASI


(cherry picked from commit e2ae18ead10d7186a76bb1a8299bfcd4ddf3820e)
(cherry picked from commit eaf8c59b3820a575742370aa6b4936d3783bc49b)
(cherry picked from commit 753fb6916ba32673a479b8c8e39cec447aa97012)

+++ test/unit/org/apache/cassandra/guardrails/GuardrailMaterializedViewsPerTableTest.java
@@ -40,13 +40,13 @@ public class GuardrailMaterializedViewsPerTableTest extends GuardrailTester
     private static final String CREATE_VIEW = "CREATE MATERIALIZED VIEW %s AS SELECT * FROM %%s " +
                                               "WHERE k is NOT NULL AND v IS NOT NULL PRIMARY KEY (v, k)";
 
-    private Long defaultMVPerTableFailureThreshold;
+    private int defaultMVPerTableFailureThreshold;
 
     @Before
     public void before()
     {
         defaultMVPerTableFailureThreshold = config().materialized_view_per_table_failure_threshold;
-        config().materialized_view_per_table_failure_threshold = 1L;
+        config().materialized_view_per_table_failure_threshold = 1;
 
         createTable(CREATE_TABLE);
     }
